+++ otp.go
@@ -17,11 +17,11 @@ type Hasher struct {
 
 type OTP struct {
 	secret string  // secret in base32 format
-	digits int     // number of integers in the OTP. Some apps expect this to be 6 digits, others support more.
+	digits int64   // number of integers in the OTP. Some apps expect this to be 6 digits, others support more.
 	hasher *Hasher // digest function to use in the HMAC (expected to be sha1)
 }
 
-func NewOTP(secret string, digits int, hasher *Hasher) OTP {
+func NewOTP(secret string, digits int64, hasher *Hasher) OTP {
 	if hasher == nil {
 		hasher = &Hasher{
 			HashName: "sha1",
@@ -39,21 +39,21 @@ func NewOTP(secret string, digits int, hasher *Hasher) OTP {
 params
     input: the HMAC counter value to use as the OTP input. Usually either the counter, or the computed integer based on the Unix timestamp
 */
-func (o *OTP) generateOTP(input int) string {
+func (o *OTP) generateOTP(input int64) string {
 	if input < 0 {
 		panic("input must be positive integer")
 	}
 	hasher := hmac.New(o.hasher.Digest, o.byteSecret())
-	hasher.Write(Itob(input))
+	hasher.Write(Itob(int(input)))
 	hmacHash := hasher.Sum(nil)
 
-	offset := int(hmacHash[len(hmacHash)-1] & 0xf)
-	code := ((int(hmacHash[offset]) & 0x7f) << 24) |
-		((int(hmacHash[offset+1] & 0xff)) << 16) |
-		((int(hmacHash[offset+2] & 0xff)) << 8) |
-		(int(hmacHash[offset+3]) & 0xff)
+	offset := int64(hmacHash[len(hmacHash)-1] & 0xf)
+	code := ((int64(hmacHash[offset]) & 0x7f) << 24) |
+		((int64(hmacHash[offset+1] & 0xff)) << 16) |
+		((int64(hmacHash[offset+2] & 0xff)) << 8) |
+		(int64(hmacHash[offset+3]) & 0xff)
 
-	code = code % int(math.Pow10(o.digits))
+	code = code % int64(math.Pow10(int(o.digits)))
 	return fmt.Sprintf(fmt.Sprintf("%%0%dd", o.digits), code)
 }
